+++ Tycoonia/Domain/Buildings/Factory/FactoryFuel.cs
@@ -13,7 +13,7 @@ namespace Tycoonia.Domain.Buildings.Factory
             };
             ProductionRate = 1m;
             EnergyConsumption = 0.5m;
-            RecipeList = new Dictionary<string, byte>
+            RecipeList = new Dictionary<string, long>
             {
                 { "Money", 10 },
                 { "Oil", 3 }
@@ -25,3 +25,4 @@ namespace Tycoonia.Domain.Buildings.Factory
         }
     }
 }
+
